+++ src/classes/chunks/CompositeDrawableEffectChunk.ts
@@ -1,0 +1,72 @@
+//
+// Imports
+//
+
+import { Chunk, ChunkOptions, ChunkParseDataOptions } from "./Chunk.js";
+
+import { Pure3DBinaryReader } from "../Pure3DBinaryReader.js";
+import { Pure3DBinaryWriter } from "../Pure3DBinaryWriter.js";
+
+//
+// Class
+//
+
+export interface CompositeDrawableEffectChunkOptions
+{
+	name : string;
+
+	isTranslucent : number;
+
+	skeletonJointId : number;
+}
+
+export class CompositeDrawableEffectChunk extends Chunk implements CompositeDrawableEffectChunkOptions
+{
+	static override parseData(options : ChunkParseDataOptions) : CompositeDrawableEffectChunkOptions
+	{
+		const binaryReader = new Pure3DBinaryReader(options.arrayBuffer, options.isLittleEndian);
+
+		const name = binaryReader.readPure3DString();
+
+		const isTranslucent = binaryReader.readUInt32();
+
+		const skeletonJointId = binaryReader.readUInt32();
+
+		return {
+			name,
+			isTranslucent,
+			skeletonJointId,
+		};
+	}
+
+	name : string;
+
+	isTranslucent : number;
+
+	skeletonJointId : number;
+
+	constructor(options : Omit<ChunkOptions, "identifier"> & CompositeDrawableEffectChunkOptions)
+	{
+		super(
+			{
+				...options,
+
+				identifier: Chunk.identifiers.COMPOSITE_DRAWABLE_EFFECT,
+			});
+
+		this.name = options.name;
+
+		this.isTranslucent = options.isTranslucent;
+
+		this.skeletonJointId = options.skeletonJointId;
+	}
+
+	override writeData(binaryWriter : Pure3DBinaryWriter) : void
+	{
+		binaryWriter.writePure3DString(this.name);
+
+		binaryWriter.writeUInt32(this.isTranslucent);
+
+		binaryWriter.writeUInt32(this.skeletonJointId);
+	}
+}