+++ CHANGELOG.md
@@ -5,6 +5,14 @@
 * Added `PackedNormalListChunk` class.
 * Added `UVListChunk` class.
 * Added `WeightListChunk` class.
+* Added `CompositeDrawableChunk` class.
+* Added `CompositeDrawableEffectChunk` class.
+* Added `CompositeDrawableEffectListChunk` class.
+* Added `CompositeDrawablePropChunk` class.
+* Added `CompositeDrawablePropListChunk` class.
+* Added `CompositeDrawableSkinChunk` class.
+* Added `CompositeDrawableSkinListChunk` class.
+* Added `CompositeDrawableSortOrderChunk` class.
 * Updated the `defaultChunkRegistry` instance with the new classes.
 
 ## 1.1.0